Refactor out BaseSpatialMeshObserver
