+++ Assets/MixedRealityToolkit/Providers/BaseSpatialMeshObserver.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cb283f9fcabe1ff46a767f251c93f24d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
